requirements.txt
python-dotenv==1.0.1
langchain-groq==0.2.2
langchain-core==0.3.28
langchain-community==0.3.13
langchain-huggingface==0.1.2
qdrant-client==1.12.1
langchain-qdrant==0.2.0
langgraph==0.2.60
pydantic==2.9.2
typing-extensions==4.12.2
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.6
